+++ src/Pages/Home/Error/Index.cshtml.cs
@@ -10,21 +10,24 @@ public class Index : PageModel
 {
     private readonly IIdentityServerInteractionService _interaction;
     private readonly IWebHostEnvironment _environment;
-        
+
     public ViewModel View { get; set; }
-        
+
     public Index(IIdentityServerInteractionService interaction, IWebHostEnvironment environment)
     {
         _interaction = interaction;
         _environment = environment;
     }
-        
+
     public async Task OnGet(string errorId)
     {
         View = new ViewModel();
 
+        CancellationTokenSource cts = new CancellationTokenSource();
+        CancellationToken ct = cts.Token;
+
         // retrieve error details from identityserver
-        var message = await _interaction.GetErrorContextAsync(errorId);
+        var message = await _interaction.GetErrorContextAsync(errorId, ct);
         if (message != null)
         {
             View.Error = message;